backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
SQLAlchemy==2.0.36
alembic==1.14.0
httpx==0.28.1
python-multipart==0.0.20
orjson==3.10.12
psycopg[binary]==3.2.4
email-validator==2.2.0
stripe==11.1.0
psycopg2-binary==2.9.9
